precision slider added in challenge settings tab
