+++ src/app/components/challenge/challengesettings/challengesettings.component.html
@@ -9,6 +9,10 @@
     </div>
     <br>
     <div class="row settings-section">
+    <span class="fw-regular col-lg-3">Precision Value:</span>
+    <mat-slider class="col-lg-9" thumbLabel [displayWith]="value" min="1.2" max="2" step="0.1" value="1.5"></mat-slider>
+    </div>
+    <div class="row settings-section">
       <div class="col-lg-3 col-md-4 col-sm-3 col-xs-3">
         <span class="fw-regular">Banned emails Ids:</span>
       </div>
