+++ src/app/components/challenge/challengesettings/challengesettings.component.scss
@@ -38,3 +38,7 @@
 .error-message {
     color: $highlight;
 }
+
+mat-slider {
+    width: 500px;
+  }
